requirements.txt
Django==4.0 
psycopg2-binary==2.9.7
djangorestframework==3.14.0
django-filter==23.2
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
django-cors-headers==4.2.0
Pillow==10.0.0
django-redis==5.3.0
 
